[Feat] 씬 체인지 패널 동기화 완료
