+++ Assets/LHW/Scene/MapLoadingSystem/LHW_MapCreateSystemTest 2.unity
@@ -548,6 +548,9 @@ GameObject:
   - component: {fileID: 152872340}
   - component: {fileID: 152872342}
   - component: {fileID: 152872341}
+  - component: {fileID: 152872345}
+  - component: {fileID: 152872344}
+  - component: {fileID: 152872343}
   m_Layer: 5
   m_Name: SceneChangePanel
   m_TagString: Untagged
@@ -612,6 +615,59 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 152872339}
   m_CullTransparentMesh: 1
+--- !u!114 &152872343
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 152872339}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e8d274f3384ba8a4aa10a491d155b412, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  roundOverTransform: {fileID: 1533162509}
+  sceneChangeInitTransform: {fileID: 1894761145}
+--- !u!114 &152872344
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 152872339}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 627855c7f81362d41938ffe0b1475957, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_SynchronizePosition: 1
+  m_SynchronizeRotation: 1
+  m_SynchronizeScale: 0
+  m_UseLocal: 1
+--- !u!114 &152872345
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 152872339}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa584fbee541324448dd18d8409c7a41, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObservedComponentsFoldoutOpen: 1
+  Group: 0
+  prefixField: -1
+  Synchronization: 3
+  OwnershipTransfer: 0
+  observableSearch: 2
+  ObservedComponents:
+  - {fileID: 152872344}
+  sceneViewId: 9
+  InstantiationId: 0
+  isRuntimeInstantiated: 0
 --- !u!1 &167608833
 GameObject:
   m_ObjectHideFlags: 0
@@ -3336,7 +3392,6 @@ MonoBehaviour:
   - {fileID: 1815037346}
   - {fileID: 0}
   losePosition: {fileID: 1756198263}
-  sceneChangeInitTransform: {fileID: 1894761145}
   roundImageShrinkDuration: 0.1
   winImageShrinkDuration: 0.1
   winImageShrinkDelay: 1.6
